+++ iGCS/Comm/CommInterface.h
@@ -16,8 +16,8 @@
 
 @required;
 // receiveBytes processes bytes forwarded from another interface
--(void)consumeData:(const uint8_t*)bytes length:(int)length;
--(void)produceData:(const uint8_t*)bytes length:(int)length;
+-(void)consumeData:(const uint8_t*)bytes length:(NSInteger)length;
+-(void)produceData:(const uint8_t*)bytes length:(NSInteger)length;
 -(void)close;
 @end
 
@@ -26,8 +26,8 @@
 // subsclasses must assign this property to use produceData
 @property (strong) CommConnectionPool *connectionPool;
 // receiveBytes processes bytes forwarded from another interface
--(void)consumeData:(const uint8_t*)bytes length:(int)length;
--(void)produceData:(const uint8_t*)bytes length:(int)length;
+-(void)consumeData:(const uint8_t*)bytes length:(NSInteger)length;
+-(void)produceData:(const uint8_t*)bytes length:(NSInteger)length;
 -(void)close;
 
 @end
